legacy/requirements.txt
beautifulsoup4==4.12.2
pdfminer==20191125
pyuca==1.2
keyboard==0.13.5
typer==0.9.0
boto3>=1.28.57
awscli>=1.29.57
botocore>=1.31.57
pydantic
json-repair==0.4.5
google-ai-generativelanguage==0.6.10
google-generativeai==0.8.3
pyaml==24.12.1
instructor==1.7.8
openai==1.58.1